chore: bulk export refactoring
